IBM compatible AT drivers housekeeping No.1: split out 486 and megapc from at.cpp
