+++ src/mame/mess.flt
@@ -87,6 +87,7 @@ argox.cpp
 asst128.cpp
 astrohome.cpp
 at.cpp
+at486.cpp
 atari400.cpp
 atarist.cpp
 atm.cpp
